Moved saveStory functionality into hook
